UI and Location Updates
Tweaked map marker click and link click behaviors. Clicking a location
now sets the zoom to 17 and centers the view on the location. Closing
the infoWindow now resets the view to it's default center and zoom
level.

Default zoom level is 14, seems to show all locations on most devices.
Location zoom level is 17, this is recommended by google. It is swell.

Updated location names an latlng's to better display the makers in
relationship to the actual locations, as well as pull the most
information possible from 4square.